smart_internal_search/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.1
django-filter==23.3
python-magic==0.4.27
python-magic-bin==0.4.14
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.3
openpyxl==3.1.2
pytesseract==0.3.10
celery==5.3.4
